% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_data_module.R
\name{choose_data_ui}
\alias{choose_data_ui}
\title{File Upload Management Module}
\usage{
choose_data_ui(id)
}
\arguments{
\item{id}{Module ID, a string that uniquely identifies the module instance.}

\item{uploaded_files}{A \code{reactiveValues} object to store all uploaded files.}

\item{pop_data}{A \code{reactiveVal} object to store population data.}

\item{curve_data}{A \code{reactiveVal} object to store curve data.}

\item{noise_data}{A \code{reactiveVal} object to store noise data.}
}
\value{
A list of reactive objects:
\itemize{
\item \code{uploaded_files}: A \code{reactiveValues} object containing uploaded file data.
\item \code{pop_data}: A \code{reactiveVal} containing population data.
\item \code{curve_data}: A \code{reactiveVal} containing curve data.
\item \code{noise_data}: A \code{reactiveVal} containing noise data.
}
}
\description{
This Shiny module manages file uploads, allowing users to select a file type,
upload a CSV or RDS file, and store the uploaded file's data in a reactive object.
It also provides feedback on the success of the upload.
}
\details{
\subsection{Features:}{
\itemize{
\item \strong{Choose file upload type}: Users can select the type of data they want to upload (e.g., Pop Data, Curve Data, Noise Data).
\item \strong{Upload CSV or RDS file}: Supports file uploads in CSV or RDS format.
\item \strong{Store uploaded data}: Data is stored in the appropriate reactive object based on the selected type.
\item \strong{Feedback on upload success}: Updates the UI dynamically to indicate successful data uploaded.
}
}
}
