% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_seroincidence_module.R
\name{estimate_seroincidence_server}
\alias{estimate_seroincidence_server}
\title{server-side computation of seroincidence estimation}
\usage{
estimate_seroincidence_server(id, pop_data, curve_data, noise_data)
}
\arguments{
\item{id}{identify a namespace}

\item{pop_data}{an object of population data}

\item{curve_data}{an object of curve data with an extra antigen column}

\item{noise_data}{an object of noise data}

\item{antigen_iso}{a vector of antigen isotype}
}
\description{
server-side computation of seroincidence estimation
}
