% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_seroincidence_module.R
\name{estimate_seroincidence_ui}
\alias{estimate_seroincidence_ui}
\title{ui for displaying to seroincidence estimation}
\usage{
estimate_seroincidence_ui(id)
}
\arguments{
\item{id}{a \code{string} to identify a namespace}
}
\description{
ui for displaying to seroincidence estimation
}
