% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_module.R
\name{import_data_server}
\alias{import_data_server}
\title{Sever-side for Data Importation}
\usage{
import_data_server(id, pop_data, curve_data, noise_data)
}
\arguments{
\item{id}{identify namespace}

\item{uploaded_files}{list of uploaded files}
}
\description{
Sever-side for Data Importation
}
