% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data_module.R
\name{inspect_data_server}
\alias{inspect_data_server}
\title{server-side data inspection}
\usage{
inspect_data_server(id, pop_data, curve_data, noise_data, imported_data)
}
\arguments{
\item{id}{identify namespace}

\item{pop_data}{population data}

\item{curve_data}{curve data}

\item{noise_data}{noise data}

\item{imported_data}{data returned by import_data_module}
}
\description{
server-side data inspection
}
