% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data_module.R
\name{inspect_data_ui}
\alias{inspect_data_ui}
\title{ui visually inspect uploaded data}
\usage{
inspect_data_ui(id)
}
\arguments{
\item{id}{a \code{string} to identify a namespace}
}
\description{
ui visually inspect uploaded data
}
