% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.serocalculator-package.R
\docType{package}
\name{shiny.serocalculator-package}
\alias{shiny.serocalculator}
\alias{shiny.serocalculator-package}
\title{shiny.serocalculator: estimates of the frequency with which seroconversions}
\description{
Translates antibody levels measured in cross-sectional population samples into estimates of the frequency with which seroconversions (infections) occur in the sampled populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ucd-serg.github.io/shiny_serocalculator/}
}

}
\author{
\strong{Maintainer}: Kristina Lai \email{kwlai@ucdavis.edu}

Authors:
\itemize{
  \item Peter Teunis \email{p.teunis@emory.edu} (Author of the method and original code.) [copyright holder]
  \item Chris Orwa
  \item Kristen Aiemjoy \email{kaiemjoy@ucdavis.edu}
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu}
}

}
\keyword{internal}
