% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tab_module.R
\name{summary_tab_server}
\alias{summary_tab_server}
\title{Server-side for Summary Page}
\usage{
summary_tab_server(id)
}
\arguments{
\item{id}{identify namespace}
}
\description{
Server-side for Summary Page
}
