% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data_module.R
\name{inspect_data_server}
\alias{inspect_data_server}
\title{server-side data inspection}
\usage{
inspect_data_server(id, pop_data, curve_data, noise_data)
}
\arguments{
\item{id}{identify namespace}

\item{value}{a continuous attribute}
}
\description{
server-side data inspection
}
